Solid Principles applied for Integration tests
